% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgarch.R
\name{plotgarch}
\alias{plotgarch}
\title{package to plot Garch model}
\usage{
plotgarch(x)
}
\arguments{
\item{x}{time series xts object}
}
\value{
plot garch(1,1)
}
\description{
This package is used to plot series using Garch(1,1) model
}
\examples{
plotgarch(x)

}
